builder_bot/requirements.txt
fastapi
uvicorn
python-dotenv
requests
tenacity
